+++ web/src/components/books/List/styles/List.tsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const List = styled.ol`
+  margin: 1em 0;
+  padding: 0;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+`;
+
+export const ListItem = styled.li`
+  list-style-type: none;
+  width: 20rem;
+`;
