+++ web/src/graphql/book.create.mutation.graphql
@@ -1,0 +1,10 @@
+mutation CreateBook($title: String!) {
+  createBook(title: $title) {
+    book {
+      ...BookDetails
+    }
+    errors {
+      message
+    }
+  }
+}
