+++ web/src/graphql/bookdetails.fragment.graphql
