+++ web/src/graphql/books.query.graphql
@@ -1,5 +1,7 @@
 query Books {
   books {
-    ...BookDetails
+    nodes {
+      ...BookDetails
+    }
   }
 }
